requirements.txt
Django==1.11
mysqlclient
Markdown
Pygments
whoosh
django-haystack
jieba
django-markdownx
django-imagekit
django-crispy-forms
pillow
django-ckeditor
django-mptt
django-notifications-hq